Allow injecting workers and WASM from outsie the modules. Draco and Basis
