+++ packages/dev/core/src/Meshes/Compression/dracoCompression.ts
@@ -6,20 +6,13 @@ import type { IDisposable, Scene } from "../../scene";
 import { Geometry } from "../geometry";
 import { VertexBuffer } from "../buffer";
 import { VertexData } from "../mesh.vertexData";
-import { DracoDecoderModule } from "draco3dgltf";
-import type { DecoderModule, DecoderBuffer, Decoder, Mesh, PointCloud, Status } from "draco3dgltf";
+import type { DecoderModule } from "draco3dgltf";
 import { Logger } from "../../Misc/logger";
+import { decodeMesh, type AttributeData, type Message, workerFunction, initializeWebWorker } from "./dracoCompressionWorker";
+import { DracoDecoderModule } from "draco3dgltf";
 
+// eslint-disable-next-line @typescript-eslint/naming-convention
 declare let DracoDecoderModule: DracoDecoderModule;
-
-interface AttributeData {
-    kind: string;
-    data: ArrayBufferView;
-    size: number;
-    byteOffset: number;
-    byteStride: number;
-    normalized: boolean;
-}
 
 interface MeshData {
     indices?: Uint16Array | Uint32Array;
@@ -27,213 +20,12 @@ interface MeshData {
     totalVertices: number;
 }
 
-interface InitDoneMessage {
-    id: "initDone";
-}
-
-interface DecodeMeshDoneMessage {
-    id: "decodeMeshDone";
-    totalVertices: number;
-}
-
-interface IndicesMessage {
-    id: "indices";
-    data: Uint16Array | Uint32Array;
-}
-
-interface AttributeMessage extends AttributeData {
-    id: "attribute";
-}
-
-type Message = InitDoneMessage | DecodeMeshDoneMessage | IndicesMessage | AttributeMessage;
-
-// WorkerGlobalScope
-declare function importScripts(...urls: string[]): void;
-declare function postMessage(message: Message, transfer?: any[]): void;
-
 function createDecoderAsync(wasmBinary?: ArrayBuffer): Promise<{ module: DecoderModule }> {
     return new Promise((resolve) => {
-        DracoDecoderModule({ wasmBinary: wasmBinary }).then((module) => {
-            resolve({ module: module });
+        DracoDecoderModule({ wasmBinary }).then((module) => {
+            resolve({ module });
         });
     });
-}
-
-function decodeMesh(
-    decoderModule: DecoderModule,
-    data: Int8Array,
-    attributes: { [kind: string]: number } | undefined,
-    onIndicesData: (indices: Uint16Array | Uint32Array) => void,
-    onAttributeData: (kind: string, data: ArrayBufferView, size: number, offset: number, stride: number, normalized: boolean) => void
-): number {
-    let decoder: Nullable<Decoder> = null;
-    let buffer: Nullable<DecoderBuffer> = null;
-    let geometry: Nullable<Mesh | PointCloud> = null;
-
-    try {
-        decoder = new decoderModule.Decoder();
-
-        buffer = new decoderModule.DecoderBuffer();
-        buffer.Init(data, data.byteLength);
-
-        let status: Status;
-        const type = decoder.GetEncodedGeometryType(buffer);
-        switch (type) {
-            case decoderModule.TRIANGULAR_MESH: {
-                const mesh = new decoderModule.Mesh();
-                status = decoder.DecodeBufferToMesh(buffer, mesh);
-                if (!status.ok() || mesh.ptr === 0) {
-                    throw new Error(status.error_msg());
-                }
-
-                const numFaces = mesh.num_faces();
-                const numIndices = numFaces * 3;
-                const byteLength = numIndices * 4;
-
-                const ptr = decoderModule._malloc(byteLength);
-                try {
-                    decoder.GetTrianglesUInt32Array(mesh, byteLength, ptr);
-                    const indices = new Uint32Array(numIndices);
-                    indices.set(new Uint32Array(decoderModule.HEAPF32.buffer, ptr, numIndices));
-                    onIndicesData(indices);
-                } finally {
-                    decoderModule._free(ptr);
-                }
-
-                geometry = mesh;
-                break;
-            }
-            case decoderModule.POINT_CLOUD: {
-                const pointCloud = new decoderModule.PointCloud();
-                status = decoder.DecodeBufferToPointCloud(buffer, pointCloud);
-                if (!status.ok() || !pointCloud.ptr) {
-                    throw new Error(status.error_msg());
-                }
-
-                geometry = pointCloud;
-                break;
-            }
-            default: {
-                throw new Error(`Invalid geometry type ${type}`);
-            }
-        }
-
-        const numPoints = geometry.num_points();
-
-        const processAttribute = (decoder: Decoder, geometry: Mesh | PointCloud, kind: string, attribute: any) => {
-            const dataType = attribute.data_type();
-            const numComponents = attribute.num_components();
-            const normalized = attribute.normalized();
-            const byteStride = attribute.byte_stride();
-            const byteOffset = attribute.byte_offset();
-
-            const dataTypeInfo = {
-                [decoderModule.DT_FLOAT32]: { typedArrayConstructor: Float32Array, heap: decoderModule.HEAPF32 },
-                [decoderModule.DT_INT8]: { typedArrayConstructor: Int8Array, heap: decoderModule.HEAP8 },
-                [decoderModule.DT_INT16]: { typedArrayConstructor: Int16Array, heap: decoderModule.HEAP16 },
-                [decoderModule.DT_INT32]: { typedArrayConstructor: Int32Array, heap: decoderModule.HEAP32 },
-                [decoderModule.DT_UINT8]: { typedArrayConstructor: Uint8Array, heap: decoderModule.HEAPU8 },
-                [decoderModule.DT_UINT16]: { typedArrayConstructor: Uint16Array, heap: decoderModule.HEAPU16 },
-                [decoderModule.DT_UINT32]: { typedArrayConstructor: Uint32Array, heap: decoderModule.HEAPU32 },
-            };
-
-            const info = dataTypeInfo[dataType];
-            if (!info) {
-                throw new Error(`Invalid data type ${dataType}`);
-            }
-
-            const numValues = numPoints * numComponents;
-            const byteLength = numValues * info.typedArrayConstructor.BYTES_PER_ELEMENT;
-
-            const ptr = decoderModule._malloc(byteLength);
-            try {
-                decoder.GetAttributeDataArrayForAllPoints(geometry, attribute, dataType, byteLength, ptr);
-                const data = new info.typedArrayConstructor(info.heap.buffer, ptr, numValues);
-                onAttributeData(kind, data.slice(), numComponents, byteOffset, byteStride, normalized);
-            } finally {
-                decoderModule._free(ptr);
-            }
-        };
-
-        if (attributes) {
-            for (const kind in attributes) {
-                const id = attributes[kind];
-                const attribute = decoder.GetAttributeByUniqueId(geometry, id);
-                processAttribute(decoder, geometry, kind, attribute);
-            }
-        } else {
-            const dracoAttributeTypes: { [kind: string]: number } = {
-                position: decoderModule.POSITION,
-                normal: decoderModule.NORMAL,
-                color: decoderModule.COLOR,
-                uv: decoderModule.TEX_COORD,
-            };
-
-            for (const kind in dracoAttributeTypes) {
-                const id = decoder.GetAttributeId(geometry, dracoAttributeTypes[kind]);
-                if (id !== -1) {
-                    const attribute = decoder.GetAttribute(geometry, id);
-                    processAttribute(decoder, geometry, kind, attribute);
-                }
-            }
-        }
-
-        return numPoints;
-    } finally {
-        if (geometry) {
-            decoderModule.destroy(geometry);
-        }
-
-        if (buffer) {
-            decoderModule.destroy(buffer);
-        }
-
-        if (decoder) {
-            decoderModule.destroy(decoder);
-        }
-    }
-}
-
-/**
- * The worker function that gets converted to a blob url to pass into a worker.
- */
-function worker(): void {
-    let decoderPromise: PromiseLike<any> | undefined;
-
-    onmessage = (event) => {
-        const message = event.data;
-        switch (message.id) {
-            case "init": {
-                const decoder = message.decoder;
-                if (decoder.url) {
-                    importScripts(decoder.url);
-                    decoderPromise = DracoDecoderModule({ wasmBinary: decoder.wasmBinary });
-                }
-                postMessage({ id: "initDone" });
-                break;
-            }
-            case "decodeMesh": {
-                if (!decoderPromise) {
-                    throw new Error("Draco decoder module is not available");
-                }
-                decoderPromise.then((decoder) => {
-                    const numPoints = decodeMesh(
-                        decoder,
-                        message.dataView,
-                        message.attributes,
-                        (indices) => {
-                            postMessage({ id: "indices", data: indices }, [indices.buffer]);
-                        },
-                        (kind, data, size, offset, stride, normalized) => {
-                            postMessage({ id: "attribute", kind, data, size, byteOffset: offset, byteStride: stride, normalized }, [data.buffer]);
-                        }
-                    );
-                    postMessage({ id: "decodeMeshDone", totalVertices: numPoints });
-                });
-                break;
-            }
-        }
-    };
 }
 
 /**
@@ -259,6 +51,27 @@ export interface IDracoCompressionConfiguration {
          */
         fallbackUrl?: string;
     };
+}
+
+/**
+ * Options for Draco compression
+ */
+export interface IDracoCompressionOptions {
+    /**
+     * The number of workers for async operations. Specify `0` to disable web workers and run synchronously in the current context.
+     */
+    numWorkers?: number;
+    /**
+     * Optional ArrayBuffer of the WebAssembly binary.
+     * If provided it will be used instead of loading the binary from wasmBinaryUrl.
+     */
+    wasmBinary?: ArrayBuffer;
+    /**
+     * Optional worker pool to use for async decoding.
+     * If provided, numWorkers will be ignored and the worker pool will be used instead.
+     * If provided the draco script will not be loaded from the DracoConfiguration.
+     */
+    workerPool?: AutoReleaseWorkerPool;
 }
 
 /**
@@ -351,67 +164,53 @@ export class DracoCompression implements IDisposable {
 
     /**
      * Constructor
-     * @param numWorkers The number of workers for async operations. Specify `0` to disable web workers and run synchronously in the current context.
+     * @param numWorkers The number of workers for async operations Or an options object. Specify `0` to disable web workers and run synchronously in the current context.
      */
-    constructor(numWorkers = DracoCompression.DefaultNumWorkers) {
-        const decoder = DracoCompression.Configuration.decoder;
+    constructor(numWorkers: number | IDracoCompressionOptions = DracoCompression.DefaultNumWorkers) {
+        // check if the decoder binary and worker pool was injected
+        // Note - it is expected that the developer checked if WebWorker, WebAssembly and the URL object are available
+        if (typeof numWorkers === "object" && numWorkers.workerPool) {
+            // set the promise accordingly
+            this._workerPoolPromise = Promise.resolve(numWorkers.workerPool);
+        } else {
+            // to avoid making big changes to the decider, if wasmBinary is provided use it in the wasmBinaryPromise
+            const wasmBinaryProvided = typeof numWorkers === "object" && numWorkers.wasmBinary;
 
-        const decoderInfo: { url: string | undefined; wasmBinaryPromise: Promise<ArrayBuffer | undefined> } =
-            decoder.wasmUrl && decoder.wasmBinaryUrl && typeof WebAssembly === "object"
-                ? {
-                      url: Tools.GetBabylonScriptURL(decoder.wasmUrl, true),
-                      wasmBinaryPromise: Tools.LoadFileAsync(Tools.GetBabylonScriptURL(decoder.wasmBinaryUrl, true)),
-                  }
-                : {
-                      url: Tools.GetBabylonScriptURL(decoder.fallbackUrl!),
-                      wasmBinaryPromise: Promise.resolve(undefined),
-                  };
+            // code maintained here for back-compat with no changes
+            const decoder = DracoCompression.Configuration.decoder;
 
-        if (numWorkers && typeof Worker === "function" && typeof URL === "function") {
-            this._workerPoolPromise = decoderInfo.wasmBinaryPromise.then((decoderWasmBinary) => {
-                const workerContent = `${decodeMesh}(${worker})()`;
-                const workerBlobUrl = URL.createObjectURL(new Blob([workerContent], { type: "application/javascript" }));
+            const decoderInfo: { url: string | undefined; wasmBinaryPromise: Promise<ArrayBuffer | undefined> } =
+                decoder.wasmUrl && decoder.wasmBinaryUrl && typeof WebAssembly === "object"
+                    ? {
+                          url: Tools.GetBabylonScriptURL(decoder.wasmUrl, true),
+                          wasmBinaryPromise: wasmBinaryProvided ? Promise.resolve(wasmBinaryProvided) : Tools.LoadFileAsync(Tools.GetBabylonScriptURL(decoder.wasmBinaryUrl, true)),
+                      }
+                    : {
+                          url: Tools.GetBabylonScriptURL(decoder.fallbackUrl!),
+                          wasmBinaryPromise: Promise.resolve(undefined),
+                      };
 
-                return new AutoReleaseWorkerPool(numWorkers, () => {
-                    return new Promise((resolve, reject) => {
+            const numberOfWorkers = typeof numWorkers === "number" ? numWorkers : numWorkers.numWorkers;
+            if (numberOfWorkers && typeof Worker === "function" && typeof URL === "function") {
+                this._workerPoolPromise = decoderInfo.wasmBinaryPromise.then((decoderWasmBinary) => {
+                    const workerContent = `${decodeMesh}(${workerFunction})()`;
+                    const workerBlobUrl = URL.createObjectURL(new Blob([workerContent], { type: "application/javascript" }));
+
+                    return new AutoReleaseWorkerPool(numberOfWorkers as number, () => {
                         const worker = new Worker(workerBlobUrl);
-                        const onError = (error: ErrorEvent) => {
-                            worker.removeEventListener("error", onError);
-                            worker.removeEventListener("message", onMessage);
-                            reject(error);
-                        };
-
-                        const onMessage = (event: MessageEvent<Message>) => {
-                            if (event.data.id === "initDone") {
-                                worker.removeEventListener("error", onError);
-                                worker.removeEventListener("message", onMessage);
-                                resolve(worker);
-                            }
-                        };
-
-                        worker.addEventListener("error", onError);
-                        worker.addEventListener("message", onMessage);
-
-                        worker.postMessage({
-                            id: "init",
-                            decoder: {
-                                url: decoderInfo.url,
-                                wasmBinary: decoderWasmBinary,
-                            },
-                        });
+                        return initializeWebWorker(worker, decoderWasmBinary!, decoderInfo.url);
                     });
                 });
-            });
-        } else {
-            this._decoderModulePromise = decoderInfo.wasmBinaryPromise.then((decoderWasmBinary) => {
-                if (!decoderInfo.url) {
-                    throw new Error("Draco decoder module is not available");
-                }
+            } else {
+                this._decoderModulePromise = decoderInfo.wasmBinaryPromise.then(async (decoderWasmBinary) => {
+                    if (!decoderInfo.url) {
+                        throw new Error("Draco decoder module is not available");
+                    }
 
-                return Tools.LoadBabylonScriptAsync(decoderInfo.url).then(() => {
-                    return createDecoderAsync(decoderWasmBinary as ArrayBuffer);
+                    await Tools.LoadBabylonScriptAsync(decoderInfo.url);
+                    return await createDecoderAsync(decoderWasmBinary as ArrayBuffer);
                 });
-            });
+            }
         }
     }
 
@@ -433,16 +232,16 @@ export class DracoCompression implements IDisposable {
      * Returns a promise that resolves when ready. Call this manually to ensure draco compression is ready before use.
      * @returns a promise that resolves when ready
      */
-    public whenReadyAsync(): Promise<void> {
+    public async whenReadyAsync(): Promise<void> {
         if (this._workerPoolPromise) {
-            return this._workerPoolPromise.then(() => {});
+            await this._workerPoolPromise;
+            return;
         }
 
         if (this._decoderModulePromise) {
-            return this._decoderModulePromise.then(() => {});
+            await this._decoderModulePromise;
+            return;
         }
-
-        return Promise.resolve();
     }
 
     private _decodeMeshAsync(
@@ -557,75 +356,67 @@ export class DracoCompression implements IDisposable {
      * @param attributes A map of attributes from vertex buffer kinds to Draco unique ids
      * @returns A promise that resolves with the decoded geometry
      */
-    public decodeMeshToGeometryAsync(name: string, scene: Scene, data: ArrayBuffer | ArrayBufferView, attributes?: { [kind: string]: number }): Promise<Geometry> {
-        return this._decodeMeshAsync(data, attributes).then((meshData) => {
-            const geometry = new Geometry(name, scene);
-
-            if (meshData.indices) {
-                geometry.setIndices(meshData.indices);
-            }
-
-            for (const attribute of meshData.attributes) {
-                geometry.setVerticesBuffer(
-                    new VertexBuffer(
-                        scene.getEngine(),
-                        attribute.data,
-                        attribute.kind,
-                        false,
-                        undefined,
-                        attribute.byteStride,
-                        undefined,
-                        attribute.byteOffset,
-                        attribute.size,
-                        undefined,
-                        attribute.normalized,
-                        true
-                    ),
-                    meshData.totalVertices
-                );
-            }
-
-            return geometry;
-        });
+    public async decodeMeshToGeometryAsync(name: string, scene: Scene, data: ArrayBuffer | ArrayBufferView, attributes?: { [kind: string]: number }): Promise<Geometry> {
+        const meshData = await this._decodeMeshAsync(data, attributes);
+        const geometry = new Geometry(name, scene);
+        if (meshData.indices) {
+            geometry.setIndices(meshData.indices);
+        }
+        for (const attribute of meshData.attributes) {
+            geometry.setVerticesBuffer(
+                new VertexBuffer(
+                    scene.getEngine(),
+                    attribute.data,
+                    attribute.kind,
+                    false,
+                    undefined,
+                    attribute.byteStride,
+                    undefined,
+                    attribute.byteOffset,
+                    attribute.size,
+                    undefined,
+                    attribute.normalized,
+                    true
+                ),
+                meshData.totalVertices
+            );
+        }
+        return geometry;
     }
 
     /** @internal */
-    public _decodeMeshToGeometryForGltfAsync(
+    public async _decodeMeshToGeometryForGltfAsync(
         name: string,
         scene: Scene,
         data: ArrayBuffer | ArrayBufferView,
         attributes: { [kind: string]: number },
         gltfNormalizedOverride: { [kind: string]: boolean }
     ): Promise<Geometry> {
-        return this._decodeMeshAsync(data, attributes, gltfNormalizedOverride).then((meshData) => {
-            const geometry = new Geometry(name, scene);
-
-            if (meshData.indices) {
-                geometry.setIndices(meshData.indices);
-            }
-
-            for (const attribute of meshData.attributes) {
-                geometry.setVerticesBuffer(
-                    new VertexBuffer(
-                        scene.getEngine(),
-                        attribute.data,
-                        attribute.kind,
-                        false,
-                        undefined,
-                        attribute.byteStride,
-                        undefined,
-                        attribute.byteOffset,
-                        attribute.size,
-                        undefined,
-                        attribute.normalized,
-                        true
-                    ),
-                    meshData.totalVertices
-                );
-            }
-
-            return geometry;
-        });
+        const meshData = await this._decodeMeshAsync(data, attributes, gltfNormalizedOverride);
+        const geometry = new Geometry(name, scene);
+        if (meshData.indices) {
+            geometry.setIndices(meshData.indices);
+        }
+        for (const attribute of meshData.attributes) {
+            geometry.setVerticesBuffer(
+                new VertexBuffer(
+                    scene.getEngine(),
+                    attribute.data,
+                    attribute.kind,
+                    false,
+                    undefined,
+                    attribute.byteStride,
+                    undefined,
+                    attribute.byteOffset,
+                    attribute.size,
+                    undefined,
+                    attribute.normalized,
+                    true
+                ),
+                meshData.totalVertices
+            );
+        }
+        return geometry;
     }
 
     /**
@@ -635,29 +426,25 @@ export class DracoCompression implements IDisposable {
      * @returns A promise that resolves with the decoded vertex data
      * @deprecated Use {@link decodeMeshToGeometryAsync} for better performance in some cases
      */
-    public decodeMeshAsync(data: ArrayBuffer | ArrayBufferView, attributes?: { [kind: string]: number }): Promise<VertexData> {
-        return this._decodeMeshAsync(data, attributes).then((meshData) => {
-            const vertexData = new VertexData();
+    public async decodeMeshAsync(data: ArrayBuffer | ArrayBufferView, attributes?: { [kind: string]: number }): Promise<VertexData> {
+        const meshData = await this._decodeMeshAsync(data, attributes);
+        const vertexData = new VertexData();
+        if (meshData.indices) {
+            vertexData.indices = meshData.indices;
+        }
+        for (const attribute of meshData.attributes) {
+            const floatData = VertexBuffer.GetFloatData(
+                attribute.data,
+                attribute.size,
+                VertexBuffer.GetDataType(attribute.data),
+                attribute.byteOffset,
+                attribute.byteStride,
+                attribute.normalized,
+                meshData.totalVertices
+            );
 
-            if (meshData.indices) {
-                vertexData.indices = meshData.indices;
-            }
-
-            for (const attribute of meshData.attributes) {
-                const floatData = VertexBuffer.GetFloatData(
-                    attribute.data,
-                    attribute.size,
-                    VertexBuffer.GetDataType(attribute.data),
-                    attribute.byteOffset,
-                    attribute.byteStride,
-                    attribute.normalized,
-                    meshData.totalVertices
-                );
-
-                vertexData.set(floatData, attribute.kind);
-            }
-
-            return vertexData;
-        });
+            vertexData.set(floatData, attribute.kind);
+        }
+        return vertexData;
     }
 }
