+++ packages/tools/babylonServer/public/basisTranscoder/1/basis_transcoder.js
@@ -19,4 +19,6 @@ if (typeof exports === 'object' && typeof module === 'object')
       define([], function() { return BASIS; });
     else if (typeof exports === 'object')
       exports["BASIS"] = BASIS;
+    else
+      globalThis["BASIS"] = BASIS;
